Fix type casting bug in feature addition / labeling

+++ deepform/data/add_features.py
@@ -71,7 +71,7 @@ def process_document_tokens(token_file, base_path):
     doc = add_base_features(doc)
 
     # Handle the features that need the whole document.
-    doc["label"] = np.zeros(len(doc))
+    doc["label"] = np.zeros(len(doc), dtype="u1")
     # The "label" column stores the TokenType that correctly labels this token.
     # By default this is 0, or "NONE".
     for feature in ["gross_amount", "flight_from"]:
